// packages/studio-base/src/panels/Button/types.ts
// This Source Code Form is subject to the terms of the Mozilla Public
// License, v2.0. If a copy of the MPL was not distributed with this
// file, You can obtain one at http://mozilla.org/MPL/2.0/

export type ButtonConfig = {
  topicName?: string;
  datatype?: string;
  value?: string;
  buttonText?: string;
  buttonTooltip?: string;
  buttonColor?: string;
};